backend/requirements.txt
# this file lists backend dependencies for local run and deployment
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.41
psycopg[binary]==3.2.9
alembic==1.16.4
pydantic==2.11.7
pydantic-settings==2.10.1
httpx==0.28.1
playwright==1.53.0
beautifulsoup4==4.13.5
lxml==6.0.0
pandas==2.3.1
numpy==2.3.1
python-slugify==8.0.4
python-dateutil==2.9.0.post0
tenacity==9.1.2
vaderSentiment==3.3.2
orjson==3.11.1
apscheduler==3.11.0
